ref(compiler): extract AssocConjSpecialization from CallSpecialization
Fifth step of splitting the CallSpecialization god-class. Move the
assoc/conj/dissoc single-call lowering and the transient assoc/conj chain
detection (plus their exclusive classifyChainHead helper) into a dedicated
AssocConjSpecialization collaborator.

Pure move: CallSpecialization dispatch and CallEmitter delegate to the new
class; the existing assoc/conj unit tests move alongside it, with an added
chain-batching test. Generated PHP is byte-identical (all integration
fixtures pass).
