+++ tests/php/Unit/Compiler/Domain/Emitter/OutputEmitter/CallSpecializationTest.php
@@ -12,8 +12,6 @@ use Phel\Compiler\Domain\Analyzer\Ast\LiteralNode;
 use Phel\Compiler\Domain\Analyzer\Ast\LocalVarNode;
 use Phel\Compiler\Domain\Analyzer\Environment\NodeEnvironment;
 use Phel\Compiler\Domain\Emitter\OutputEmitter\CallSpecialization;
-use Phel\Lang\Collections\Map\PersistentMapInterface;
-use Phel\Lang\Collections\Vector\PersistentVectorInterface;
 use Phel\Lang\Keyword;
 use Phel\Lang\Symbol;
 use Phel\Shared\CompilerConstants;
@@ -21,77 +19,6 @@ use PHPUnit\Framework\TestCase;
 
 final class CallSpecializationTest extends TestCase
 {
-    public function test_assoc_on_typed_map_specialises_to_put(): void
-    {
-        $env = $this->env();
-        $node = $this->coreCall('assoc', [
-            $this->localWithTag('m', PersistentMapInterface::class),
-            new LiteralNode($env, 'k'),
-            new LiteralNode($env, 1),
-        ]);
-
-        self::assertSame('put', CallSpecialization::typedAssocConjDissocMethod($node));
-    }
-
-    public function test_assoc_on_typed_vector_specialises_to_update(): void
-    {
-        $env = $this->env();
-        $node = $this->coreCall('assoc', [
-            $this->vectorLocal('v'),
-            new LiteralNode($env, 0),
-            new LiteralNode($env, 42),
-        ]);
-
-        self::assertSame('update', CallSpecialization::typedAssocConjDissocMethod($node));
-    }
-
-    public function test_conj_on_typed_vector_specialises_to_append(): void
-    {
-        $env = $this->env();
-        $node = $this->coreCall('conj', [
-            $this->vectorLocal('v'),
-            new LiteralNode($env, 99),
-        ]);
-
-        self::assertSame('append', CallSpecialization::typedAssocConjDissocMethod($node));
-    }
-
-    public function test_dissoc_on_typed_map_specialises_to_remove(): void
-    {
-        $env = $this->env();
-        $node = $this->coreCall('dissoc', [
-            $this->localWithTag('m', PersistentMapInterface::class),
-            new LiteralNode($env, 'k'),
-        ]);
-
-        self::assertSame('remove', CallSpecialization::typedAssocConjDissocMethod($node));
-    }
-
-    public function test_variadic_assoc_is_not_specialised(): void
-    {
-        $env = $this->env();
-        $node = $this->coreCall('assoc', [
-            $this->localWithTag('m', PersistentMapInterface::class),
-            new LiteralNode($env, 'k1'),
-            new LiteralNode($env, 1),
-            new LiteralNode($env, 'k2'),
-            new LiteralNode($env, 2),
-        ]);
-
-        self::assertNull(CallSpecialization::typedAssocConjDissocMethod($node));
-    }
-
-    public function test_conj_on_untyped_local_falls_back(): void
-    {
-        $env = $this->env();
-        $node = $this->coreCall('conj', [
-            new LocalVarNode($env, Symbol::create('v')),
-            new LiteralNode($env, 99),
-        ]);
-
-        self::assertNull(CallSpecialization::typedAssocConjDissocMethod($node));
-    }
-
     public function test_not_eq_peephole_fires_when_inner_eq_is_typed(): void
     {
         $this->env();
@@ -226,11 +153,6 @@ final class CallSpecializationTest extends TestCase
             ),
             $args,
         );
-    }
-
-    private function vectorLocal(string $name): LocalVarNode
-    {
-        return $this->localWithTag($name, PersistentVectorInterface::class);
     }
 
     private function localWithTag(string $name, string $tag): LocalVarNode
